fixed some errors

+++ AMFICOM/v1/database/schema/mapview/object/mapview.sql
@@ -8,10 +8,10 @@ CREATE TABLE MapView (
  domain_id VARCHAR2(32),
  name VARCHAR2(128),
  description VARCHAR2(256),
- longitude FLOAT,
- latitude FLOAT,
- scale FLOAT,
- defaultScale FLOAT,
+ longitude NUMBER(12, 6),
+ latitude NUMBER(12, 6),
+ scale NUMBER(12, 6),
+ defaultScale NUMBER(12, 6),
  map_id VARCHAR2(32), 
 --
  CONSTRAINT mv_pk PRIMARY KEY (id),
@@ -23,7 +23,6 @@ CREATE TABLE MapView (
   REFERENCES Domain (id) ON DELETE CASCADE,
  CONSTRAINT mv_map_fk FOREIGN KEY (map_id)
   REFERENCES Map (id) ON DELETE CASCADE
-
 );
 
 CREATE SEQUENCE mapview_seq ORDER;
